fix: 🐛 refactor immutable package made compilation error test
